+++ CHANGELOG.md
@@ -13,6 +13,14 @@ If you used custom channels, you must updated your config when updating from 1.1
 * In work: improve handling of channel names: characters not allowed in HomeKit are now replaced with whitespace
 
 
+## 1.1.7-beta.1 (2021-04-16)
+* Fixed issue where channel numbers would appear multiple times
+* Fixed issue where user-configured channel names would be discarded
+* Fixed issue where display of recordingState in logs was incorrect
+* Improved consistency of log display
+* Added support of DisplayOrder for proper channel sorting
+
+
 ## 1.1.6 (2021-04-14)
 * Fixed issue where showChannelNumbers was not working
 * Fixed issue in config.schema where channelNames was shown in the wrong config location